input: do not merge identical steps with different sandboxes
When the sandboxInvariant policy is enabled the sandbox is considered
an invariant of the build. Given an identical package that is
referenced with different sandboxes in the recipes we previously
treated them as fully identical packages. But this might have prevented
the user from selecting between the desired instance with a concrete
sandbox. Instead the user could experience that he chose a path that
should have used a particular sandbox but when building actually
another sandbox was used. While the *result* should be identical we
must still honor the users choice which sandbox to use.