input: prevent false reuse of packages
Packages in the package graph are reused if they are considered to yield
the identical result. This is determined by comparing the input of a
recipe (environment, tools, sandbox) and see if the same input was
already used to compute a package. This algorithm relies on the property
that *all* relevant input variables are examined.

Before this commit that was not true for all properties of tools and
sandboxes. Only the most important property of a tool/sandbox was
considrered: its variant-id. But there are other important properties as
well, e.g. attached environment variables, netAccess etc. These
properties were not considered and could thus lead to reuse of packages
that are different because of these tool/sandbox properties.

This commit attaches a resultId to each CoreTool and CoreSandbox that
ought to cover *all* properties of the tool/sandbox. This resultId is
then used in the package resultId calculation and the package matcher to
find reusable packages.