invoker: run commands with step environment by default
Before Bob 0.17 all commands, including checkout asserts and SCMs, were
compiled into one big bash script and executed in the same environment.
Starting with the multi language support the SCMs are directly invoked
by Bob. This lead to a regression where these commands would not get the
checkoutVars added to their environment any more.

Change the invoker to always apply the step environment. This is only
disabled for the fingerprint scripts (that have their own
'fingerprintVars' setting) and the checkout/build/packageScript itself
because the language handler already takes care of it.

Fixes #355.