archive: expand 'clean' expression to limit artifacts
The retention expression is expanded to something loosely resembling
SQL. An optional 'LIMIT <num>' after the predicate can limit the number
of retained artifacts. By default the artifacts are sorted by build date
in descending order so that the most recent <num> builds are kept. A
different field may be used with 'ORDER BY <field>'. The ordering can be
specified by 'DESC' (descending) or 'ASC' (ascending).

This is the first step to solve #351.