input: whitelist env variables consistently on Windows
There is no reason to make a distinction between native Windows and MSYS
in general. Even if Bob is called from an MSYS environment it is very
likely used to call native Windows executables. In this case the regular
environment variables should be made available because applications can
rely on them to be available.

Relates to BobBuildTool/basement#22.