git: add submodules support
Submodules can now be cloned if requested. Recursive submodule clones
are supported too. By default they are cloned shallowly but this can be
overridden.

The update of submodules is implemented with a dedicated logic for main
modules that are cloned from a branch. A vanilla 'git submodules update'
does always reset the submodule to the commit that is tracked in the
tree. This will revert any user changes which is not what we want.
Instead Bob will first determine which submodules are unmodified
(detached HEAD and on correct commit) and only update those modules.

The status and audit code is adjusted accordingly. Any changes in
submodules and sub-submodules are already detected as modification.
Additionally any missing submodules or unexpected submodules (submodule
checked out even though 'submodules' or 'recurseSubmodules' is not
enabled) are flagged as modification. This is not detected by a plain
'git status'.

Fixes #353.