input: introduce noUndefinedTools policy
It has been historically permissive to use tools in a recipe that are
not defined. Such declarations were silently ignored and could thus
lead to build errors. Because it is always possible to define a dummy
tool with the required name there is no good reason for this behavior.
The noUndefinedTools policy rectifies that. When the policy is on the
new behavior any undefined tools that are used lead to a parsing error.