input: add {checkout,build,package}ToolsWeak
Some tools do not make a difference in the result of a package despite
their presence. These are typically script interpreters. How the
interpreter is built is usually without influence on the executed
scripts. Such tools may be included as weak tools.

Weak tools are just identified through their name at the using package.
If a weak tool is renamed the dependent packages will still be rebuilt.
Any other changes to weak tools do not cause new artifacts to be built.