scm: add base methods for inline switch of SCMs
If the digest-script of a SCM changes it would normally moved to the
attic and a new checkout would be done. This commit adds the basic
methods to switch a SCM to the new specification without doing a fresh
checkout.

The schema uses two steps: a canSwitch(() method that checks if an
inline switch is possible and the switch() method that does the actual
switch. If either function returns False or if switch() fails with an
exception the SCM will moved to the attic and a fresh checkout is done.
In any case the invoke() method will be called afterwards.