git: support inline switch of ref/url
The change of the URL is supported by optimistically fetching the new
remote and using it. If the new repository does not share a common
history the fast-forward merge will fail and the attic logic will
replace the old module. The only drawback is that the SCM will be dirty
in the attic because of the half done checkout and the switched remote.

Branch/tag/commit changes are done straight forward. The only constraint
is that submodules are *not* used. Moving or removing submodules is not
well supported by us.