input: remove '#<num>' from SCM 'recipe' property
The trailing number was intended to signify the number of the
checkoutSCM entry in the source file when the query-scm command was
first created. This has never worked in the face of recipes that use
multiPackage. Just remove this artefact and hope that nobody has used
it yet.

Being a plain file name it is now possible to match it with
scmOverrides. This comes in handy if only a particular file should be
overridden.