state: avoid fsync
Call fsync() only once when Bob commits the state changes. At runtime
the new state is written into a separate file with a checksum. When
terminating, the file is fsync()ed and replaces the old state. In case
of a crash the new state will be picked up the next time if the checksum
is ok. This should detect and reject partially written files in case of
a machine crash.

Fixes #365.