input: implement setup scripts
Setup scripts are prepended to the regular scripts. Defining a setup
script does not yet enable the corresponding step. Conceptually a setup
script is intended to define helper functions or variables but they
should not yet execute anything. They are included when entering the
shell environment of a step (i.e. calling `build.sh shell`). As such
they are intended mainly for classes so that the definitions of a class
are automatically available in the shell environment.

Fixes #37.