Fix missing checkouts on Jenkins
With the introduction of setupScripts in d4a6cf4b the getScript() method
started to return None if no shell script is run. Unfortunately the
Jenkins logic wrongly used this value to determine if some shell must be
invoked. In case of a single url-SCM this was now wrong because the "bob
_invoke" must be run but getScript() returns None because there is no
additional real shell script.

Just run "bob _invoke" unconditionally from now on. Also fix the various
helpers to always return a string where the caller expects one.

Fixes #384.