input: fix graph cycles on weak tools
With the introduction of weak tools the build graph could get cycles
where the recipes and their package description are actually cycle free.
This can happen because the variant-id of a used weak tool is actually
not significant, causing distinct nodes to be collapsed into a single
one. The build logic will then get stuck because it assumes that the
graph is acyclic.

To work around this problem the optimizations enabled by weak tools are
partially reverted. Weak tools are only considered for the build-id
calculation and now let the variant-id untouched. This restores the
build behaviour that was present before weak tools were introduced. But
for binary artifacts weak tools are still considered. This keeps
compatibility with previously built artifacts and will still match
artifacts if the weak tool changes.

The drawback is that changes to weakly used tools will lead to rebuilds
of all using packages. Because the build-id does consider weak tools
this will most likely immediately re-download the same package again.
This might be confusing to the user.

Fixes #387.