bob: add notion of a "storage path"
The storage path is where the files of the step are stored. For checkout
and build steps this is always the workspace path. But package steps can
be shared globally and thus the directory may lie outside of the project
directory.

Even if the final storage location (after the package was built) is
somewhere else, the workspace path is kept. It serves as identifier in
numerous locations of the build logic and is also relevant to the user
to "find" the package.