share: add core shared package handler
Shared packages are stored similar to the existing Jenkins
implementation. Additionally there is some meta data for the whole
repository and per-package. This is mainly used for garbage collection
but also when using a package.

All operations should be safe against concurrent usage. Multiple
instances can install the same package. A garbage collection can run too
concurrently. This is solved by advisory locking of the meta data. Locks
are only held shortly and longer operations are done in temporary
directories.