build: fix location tracking of non-relocatable packages
Starting with 0d93b3c the exec path might be relative to the project
root directory. Unfortunately this breaks the location tracking of
non-relocatable packages which relied on the property that the exec path
is always absolute. Restore the old behaviour by ensuring the path is
always converted to an absolute path.

Fixes #410.