input: check bobMinimumVersion before config.yaml schema
Newer versions of Bob might add new keys to config.yaml. A proper
config.yaml will set "bobMinimumVersion" to the appropriate version but
it requires us to check the version *before* the schema is checked.
Otherwise the newer keys will be rejected first which leads to confusing
error messages.

Fixes #420.