git: make shallow clones with explicit commits more reliable
On a shallow clone the fetched tip(s) are only partially cloned. If the
scm is configured to use a specific commit it might be possible that the
commit was not downloaded if the branch advanced too much. Add a special
fallback to re-fetch from the server with the explicit commit id if the
initial fetch failed to download the required commit.

Note that we cannot just always request the commit explicitly when
fetching. Servers are usually configured to *not* hand out arbitrary
commits.