workflow: use shallow clone
The setuptools_scm module can cope with the fact that no version can be
determined and falls back to 0.1. In case of a tag-build the right
version is visible and PyPI packages are built with the correct version
number.

For CI builds the exact version number is also not critical.
bob.develop.version.getVersion() will fall back to the hard coded minor
version. This is sufficient for the tests as it is the base version of
the current development cycle.