invoker: do not fail on environment variable collisions
On Windows environment variables are case-insensitive but
case-preserving. Unfortunately MSYS2 creates variables that only differ
in case (e.g. TMP vs. tmp). Remove the hard failure from the check and
convert it into a warning because Bob versions before 0.17 worked
without failure in this case. Also fix the code to catch only
collisions that are created by variables defined in the recipe itself.
Duplicate keys in the inherited environment are not reported.

Fixes #432.