bob: add out-of-tree build support
Adds a new "init" command which initializes a build tree that points to
the project root directory where the recipes are read from. The build
tree might have it's own default.yaml too.

Technically the build tree has a special file ".bob-project" that holds
the path to the project root directory. If not present it defaults to
the current working directory as before.

Fixes #435.