jenkins: always obey sandbox when calculating job graph
In contrast to the local build logic the Jenkins job generator has to
eagerly create the whole job graph. In conjunction with the
sandboxInvariant policy this might lead to cyclic jobs because the
sandbox was not taken into account. If the same artifact (as in
variant-id) is built inside *and* outside of the sanbox in the same
project then a cycle was created.

The Jenkins logic will now always take the configured sandbox into
account. Steps with the same variant-id but different sandboxes are
built separately. This might lead to duplicated checkout- and
build-steps but is the simplest bullet-proof solution to the problem.

Fixes #438.