input: add checkoutDep dependency flag
Normally dependencies are only available to the buildScript. But
sometimes it is necessary to access dependencies in the checkoutScript
already. Currently this requires to copy the sources in the buildScript
where access to all dependencies is possible.

The newly added checkoutDep flag of dependencies allows to make
selected dependencies available for the checkoutScript. The buildScript
will still retain access to the depency as always. Provided
dependencies of such packages are *not* implicitly made available. The
number of checkout dependencies should be kept as small as possible to
prevent accidental creation of checkout variants, which are usually
confusing to users.