input: cache sandbox when reconstructing graph
The CoreRef.refDeref() method needs to reconstruct the concrete
Step/Package from the reduced core graph. There is a cache to prevent
unneeded dereferencing of already seen edges. Unfortunately the cache
was not filled for sandbox references. Depending on the complexity of
the sandbox partial graph this could drastically increase the time to
traverse the graph.