bob: hash bob sources platform independent
When mixing platforms, e.g. configuring a Jenkins from Linux that is
running on Windows, the source hash must be identical. The hashDirectoy
function is not usable because it makes a distinction based on the
platform in the expectation that such things like soft links might be
different. As far as the Jenkins is concerned we are just interested in
regular python files.