jenkins: add host-platform
Allows distinction between native Windows and MSYS2. By default the
Jenkins platform is assumed to be the same as where Bob is executed.
Because the host platform is now an enumeration the old '-w' option is
deprecated. It will still be understood and existing projects will keep
on working too.