bob: introduce intermediate representation
The intermediate representation is a facade the basically provides the
same methods as the input module of parsed recipes/packages/steps. The
intermediate representation will take over all methods that deal with
actual workspace paths and the like. This leaves the "input" module only
with the pure recipe information, independent of the actual build
backend.

The intermediate representation is still abstract. It provides the
common behaviour but not how the data is actually stored. This is left
to the backend itself.

Only required information is stored. This means that for mere
dependencies, only a subset of all available fields is stored. Usually
transitive dependencies are not stored at all. This is done to minimize
the information in the intermediate representation which is crucial if
it must be serialized, e.g. on Jenkins.