builder: do not check for collisions on Jenkins
On Jenkins the SCMs run before any other build step. This means that the
directories of git and svn repositories will always exist when the
checkout step logic is run. The check is even not necessary because a
change in the checkout step will always result in a new workspace path
being allocated, so there is no risk to checkout to old clones.