jenkins: build via intermediate representation
Use the same build logic as local builds on Jenkins. For that the
relevant build data is extracted for each job and included as compressed
json.

Because there must always be a binary artifact for each package that is
built in a job, a internal archiver is added. This archiver will write
the build-id file and, if required, create the artifact tgz. It will
also be used to extract all dependencies from upstream jobs.