state: atomically replace uncommitted state
Saving the current state might fail due to various reasons, e.g. a full
disk. Depending on the runtime there might already exist an uncommitted
state file. Whenever we write a new state we must make sure to first
write the full state, close the file and then replace this as the new
uncommitted state.

Previously the half written state would be blindly committed in the
assumption that it was already fully written. The reason is that the
finalize() method is always called, even if the error originated from
the __save() method itself. This left behind a wrecked project state.
Also overwriting the uncommitted state directly is not a good idea
because the previously uncommitted state is definitely lost.

The state might still get corrupted if the machine crashes because the
final fsync() is only called when Bob exits. The existing checksum
mechanism should catch this and fall back to the last known good state.

Fixes #470.