Make file replacement more robust on Windows
It appears that os.replace() can spuriously fail with a PermissionError
on Windows. This seems to be caused by other processes (Explorer?, anti
virus?) that have opened one of the files concurrently. Add a central
workaround function and use it consistently. It will retry with a
back-off timer which will hopefully be sufficient to mask the problem.