builder: make sure checkout audit trail is generated
If the generation of the audit trail for the checkout step fails, it
needs to be tried again on the next run. Otherwise we will end up
executing the next steps without input audit trail and thus wrecking all
further audit trail generation.

Fix it by setting the stored result hash of the source workspace only
after the audit trail was generated. This will ensure its generation on
the next run. The removed comment was incorrect since 2c23aeab which
removed the subtle dependency of the audit trail to the result hash.