spec: only add additional sandbox mounts from same package
If a package is built in the sandbox it is defined that all previous
steps of the package are available in the sandbox. The code that added
these extra dependencies for the sandbox relied on the fact that
checkout steps have no dependencies. With the introduction of the
checkoutDep flag this is no longer the case. It is thus required to stop
iterating steps once we reached the checkout step.

Fixes #479.