git: improve workflow with commits + branches
This changes the behavior when working with recipes, where a commit and
a branch is specified. Historically Bob simply gave priority to the
commit or tag and moved to a detached HEAD state. Now, controlled under
the "gitCommitOnBranch" policy, Bob will check if the commit is on the
specified branch or fail the build otherwise. Additionally the branch is
checked out at precisely the given commit. This avoids the `detached`
HEAD state .

Co-authored-by: Jan Klötzke <jan@kloetzke.net>