Work around Windows WSL bash issues
We need to work around a weird behaviour on Windows when WSL is enabled
but no distribution is installed. In this case the subprocess module
(which internally uses CreateProcess()) cannot execute bash even though
it's in %PATH%. Interrestingly cmd.com and powershell.exe look at %PATH%
first and can execute bash successfully.