jenkins: use canary to detect workspace creation
If an SCM is used that is supported by Jenkins natively, the directory
will always be present, even if if was deleted previously. Use the
canary as indicator if the workspace was (re-)created in this case.
Previously the created workspace was not detected, leading to missing
audit trails and potentially wrong checkouts because the checkoutScript
is not run again.