stringparser: add get-tool-env function
Query defined variables of a tool. The tool must only be defined and is
not required to be used by the recipe. An optional default value can be
passed that will be used if the variable is not defined in the tool.
Otherwise undefined variables will fail the parsing.

Fixes #486.