build: add checkout update facility in build-only mode
By default no checkout scripts are run when building with --build-only.
Some use cases practically require the checkoutScript to be always run,
through. A typical example are code generators that generate sources
from some high level description. These generators must be run every
time when the user has changed the input. A recipe or class can
explicitly opt in to run their checkoutScript also in build-only mode to
cover such a use case. This is done by either setting checkoutUpdateIf
to True or by a boolean expression that is evaluated to True. Otherwise
the checkoutScript is ignored even if some other class enables its
script. The checkoutUpdateIf property thus only applies to the
corresponding checkoutScript in the same recipe/class.

The build logic already did support the import SCM which was run in
build-only mode too. Unify the handling of the import SCM with the
selective enabling of checkoutScript's on build-only builds.

Fixes #452.