jenkins: fix build for packages with invalid build script
A recipe might declare dependencies but not actually use them if it has
no build script. While there is still a dependency on package level, it
will not be built and there won't be an upstream job that builds the
unused dependency. This must be taken into account when reading the
build-ids of all upstream jobs. Make sure to skip invalid build- or
checkout-steps which might have dependencies but which are not built and
thus unavailable.