git: status: avoid tags been shown as unpushed
It's not uncommon to have tags which do not belong to a branch. The current
behavior of the status command is to show them as 'unpushed' even if they exist
as a remote tag and nothing was changed locally.
Exclude all tags from the 'unpushed' - status output to avoid this.