git: do not clean checkout on Jenkins
Historically (since 84b9d6a2) Bob enabled the feature to call "git
clean" to remove untracked files from the checkout. When building
locally this is never done. The difference was not visible until lately
because Bob did always run the checkout script if a job was scheduled.
Since Bob will now skip checkout steps if nothing has changed, this will
break the build for recipes that generate files in the git repository in
the checkout script.

Simply remove the "git clean" feature on Jenkins. There is always the
--clean switch for recipes that cannot cope with checkout scripts that
are run more than once.