workflow: don't build binary wheel on Windows
The wheel that is built on Windows is lacking the bob-namespace-sandbox
and has no other binary artifacts either. This will thus create a
generic Python 3 wheel that can be installed on any Python 3
installation, including Linux. So if none of the Linux builds matches,
e.g. due to different libc, this will still install the generic wheel
and create a broken installation because bob-namespace-sandbox is
missing.

Fixes #502.