build: warn if files collide on --destination
When building with --destination the result of all built packages is
blindly copied to the desired directory. It has never been an error if
files are overwritten by that to support building multiple times in a
row with the same destination. But if multiple built packages contain
the same files then they silently overwrite each other. Add a warning
message for this case because the destination is in an undefined state.