project-generators: enable use of package queries
While many of the generators do not need it there are use cases where a
generator needs to handle a pathspec rather than a package-path. To
support this a projectGenerator can set the additional `query`
property to `True`. In this case the first argument for the generator
are the package objects returned by `queryPackagePath`.