url: use python extractor fallbacks only on win32
The tarfile and zipfile Python modules are slower and have behavioural
differences compared to their native implementations. The tarfile module
in particular differs in the handling of file modes (does not behave
like "--no-same-permissions"). The zipfile module looks safer but is
certainly slower than the native implementation.

As permissions are not an issue on Windows, keep the Python fallbacks
at least on native Windows.