import: fix digest script
The asDigestScript() method of import SCMs erroneously did not include
the 'dir' attribute. This cannot easily be fixed because it will
directly affect the variant-id and thus cause rebuilds and invalidate
existing binary artifacts. Instead, the fixImportScmVariant policy is
introduced to enable fixing the bug for new projects without breaking
old ones.