build: track state for build-only checkout step updates
The tracking of recipe changes was not sufficient for build-only updates
of checkouts that involve the import-SCM or recipes that utilize
checkoutUpdateIf. We were over-cautious and prevented running these
updates if *anything* related to the checkout step changed.

To make Bobs behaviour more predictable, track the related state more
closely. Updates of checkouts in build-only mode are now only prevented
if an import-SCM is changed. On the other side, changes to the
checkoutScript will trigger the update in build-only mode even if the
script is deterministic.

Fixes #506.