builder: fix garbling of shared packages
Packages at shared locations are immutable. If the recipe was changed
and the package is unshared, the workspace link in the project tree is
removed. Problem is, the os.path.isdir() function returns True for
symlinks that point to directories. This actually then clears the shared
location instead of removing the link. Fix it by reversing the
condition.

Fixes #522.