git: fix scm switch when going back to older revision
If gitCommitOnBranch is active and commit and branch are configured in
the recipe changing the commit back to a older one did not update the
sources back to the configured version. Instead they remained on the
current version.

To fix this use `git reset --keep` instead of `git merge --ff-only` when
updating, but make sure we do not lose any commits.