git: only change the workspace if really necessary
There is no need to update a workspace if it's already on the correct
tag / commit. In case a repo is moved to a new commit/tag and this new
commit/tag was written to the recipe, the previous behavior was to
switch to the new commit running the checkout again including all
checks.

In pratice there are cases where the new commit hasn't been pushed to
the configured branch when running bob again. The commitOnBranch check
noticed that the commit did not match the branch and the build fails. To
recover from this a developer needs to either move it's stuff back from
the attic, write a scmOverride, ... which is annoying and not really
necessary as the workspace is already on the correct commit.

To make this more user friendly relax the "Branch ... did not contain
.." error if the workspace is on the configured commit and leave the
workspace untouched.