Support nested SCMs
Add support to checkout an SCM into a subdirectory of another SCM. This
requires that the SCM with the upper directory is first in the list
before any other SCMs in subdirectories. Additionally, SCMs that are
natively supported by Jenkins plugins (git, svn), cannot be nested into
the other SCMs (cvs, import, url).  The reason is that Jenkins SCM
plugins always execute before anything else in a Jenkins job.

This relaxes the rules that did not allow any nesting before. There are
some drawbacks of such setups. Namely, if an SCM is moved to the attic,
all their nested SCMs are affected too.