builder: checkout modifications trigger rebuild
If there is a checkoutScript or some checkoutAssert's and the checkout
workspace was changed since the last time, we must re-run the step.

This fixes a problem on Jenkins builds where the checkoutScript modifies
a git repository. This change will be squashed by the Jenkins git plugin
every time the job runs. Now if the script was marked deterministic, Bob
still has to run it again to re-apply any modifications.

Another example are code generators that should probably be run again if
the workspace has changed. In fact, if the script was enabled for
checkout-only updates (checkoutUpdateIf), then such modifications would
already trigger the re-run if the workspace was modified.