builder: calculate build-id before executing step
Calculating the build-id after a step was executed has the unpleasant
effect that the job slot must be briefly relinquished. When running
multiple jobs in parallel, this will give the others an opportunity to
take over the job slot. In the worst case some other, long running task
is started. The original job is not finished, though, until the audit
was created.

When calculating the build-id before execution, this cannot happen. The
audit generation will run uninterrupted with the current job slot until
the end.