builder: do not ignore cancellation in build-id calculation
If the user cancels the build, we have to exit as fast as possible.
Historically the build-id calculation was shielded (see 8cc700319e)
because we did not want to re-execute finished builds that just needed
the build-id calculation to be done.

Now that the build-id is calculated before the step is executed, this
problem does not exist any more. On the contrary, this shielding
actually prevents cancellation and new jobs are still scheduled. Just
remove the shielding and things behave as expected.