url: regain mtime when copying files
The mtime itself is irrelevant for the workspace hash. But if a file is
copied from a local mirror we don't want to get the file a newer mtime
than the source mirror file. Otherwise the mirror will be updated with
the same file again. Likewise, the mirror should get the same file
attributes like the fetched file.