url: add defaultFileMode policy
Historically the URL SCM applies a file mode of 0600 (user
read/write only) to all files that are fetched via HTTP(S) or FTP. For
locally copied files (``file://`` URLs or bare file names) though, the
file mode of the source file is retained. This is both inconsistent and
not predictable for local file because the result will depend on the
source file.

The defaultFileMode policy will enforce a mode of 0600 to all files,
regardless of the origin. This will prevent any accidental checkout
inconsistencies in the future.