archive: replace live-build-id and fingerprint caches
These caching data structures are prone to incorrect data e.g., if some
checkout is not really deterministic. The old behaviour was to never
replace the files which sets the very first execution result into stone.
If that was incorrect, there are no tools to purge the wrong cache
except deleting the whole binary archive.

Instead, Bob now always uploads the live-build-id and fingerprint
caches. These files are small so there should be no performance impact.
But at least the cache will reflect the last build, giving the user the
possibility to replace the wrong predictions with the right ones.