jenkins: add option to always checkout
The Jenkins build is usually used to provide binary artifacts. This
implies that the build should be as stable as possible. If a checkout
step of some recipe is not fully deterministic, though, then relying on
live-build-ids may never rectify the problem. It might also hide such a
problem only until the checkout is really forced for some other reason.

To increase the robustness, add a "scm.always-checkout" extended option
that defaults to "true". If enabled, checkout steps will always be
executed and live-build-ids will only be uploaded but never used. If the
user wants to rely on live-build-ids to speed up builds he can disable
the scm.always-checkout option.