input: don't coalesce potentially different fingerprints
If a package is fingerprinted, the result is effectively declared to
depend on the host system. How exactly is not known to Bob. Despite the
actual host system, each sandbox represents a potentially different host
system environment. This means that the same fingerprint script might
yield different results if being executed in a sandbox, a different
sandbox or the host system.

This implies that packages which are fingerprinted need to be built
separately for the host system and each sandbox image. With the
introduction of the sandboxInvariant Bob erroneously coalesced
fingerprinted packages which only differ in their host/sandbox
environment. Now Bob will always use separate workspaces again for such
packages.

The fix will impact the calculation of the variant-id which might lead
to some rebuilds. On the other hand, the build-id is not affected and
existing binary artifacts will continue to be used.