Let sandbox keep current user identity
Apparently some license check tools require to observe the same user and
group ID inside the sandbox. Add an optional choice to the sandbox
definition and user configuration to override the user inside the
sandbox. The default is "nobody" which reflects the traditional behaviour.
The new options "root" and "$USER" set the UID to either 0 (root) or keep
the UID of the current user.

Co-authored-by: Jan Klötzke <jan@kloetzke.net>