input: add 'inherit' dependency property
This property controls the inheritance of environment, tools and sandbox
for a dependency. It defaults to true so everything is inherited.

If set to `false` environment, tools and sandbox are dropped. The
dependency starts with the default-environment, no tools and no sandbox.

This becomes particularly useful when an existing root-package should
become a dependency of another root-package, e.g. for building an
installer.