builder: consider all execution paths
The builds step implicitly depends on the execution path of all
dependencies. So far we only tracked arguments. But some build systems
also store the absolute path to tools. In this case, the build will fail
if a tool path changes.

The simple approach by using getAllDepSteps() will additionally add the
sandbox path too. This will lead to rebuilds if the sandbox image
changes. This should not happen too often and is probably not even a bad
idea.