input: evaluate deps condition first
If a dependency is disabled, all other variable substitutions in this
dependency shall be skipped. That has been the case already, except the
dependency name itself. Unfortunately, just moving the condition check
to the front is not sufficient because it is allowed to name disabled
dependencies in provideDeps. So we do the name substitution anyway but
will ignore a failure if the dependency shall be skipped.