input: move tool fingerprint mask into steps
To calculate the variant-id, the applicable tools for each step must be
known. The variant-id of a step is calculated already in its
constructor. So far this worked because the list of used tools was
static for each recipe. Now that we want to make tool usage conditional,
we have to pass the tool mask directly into each step.