stringparser: skip over unused substitution parts
Variable substitutions can have a default value ("${VAR-default}") or
an alternate value ("${VAR+alternate}"). The "default" and "alternate"
parts can itself be comprised of variable/function substitutions.

Now, if the default/alternate part is unused, it should be skipped
altogether. So far we always substituted them, even if they were
unused. This is undesired because it causes constructs like

  ${VAR:+${VAR}}

to fail which contradicts POSIX shell substitution behaviour which we
try to follow. Rectify this by suppressing substitution in unused parts
and just skip them.