build: add bundle build
A bob-bundle collects all scm's neccessary to build a package in a
tar-file. Next to the tar-file a scmOverrides file is generated to be
able to rebuild the package only with the bundle.

Such a bundle can be used to build on a air-gapped system or to archive
all sources of a build.

ATM only bundling of git and url-scm's is supported.