layers: internalize config.yaml handling for layers
The handling of layers just depends on their config.yaml. Actually, only
a subset of the project configuration is needed. Instead of using the
regular RecipeSet parsing function with an early exit, do all
config.yaml processing in the layers module itself.

This fixes a couple of subtle corner cases that so far got unnoticed:

 * We still parsed the system- and user-global default.yaml
   configuration files. The data from these files was not used except
   the URM SCM mirrors. Now everything depends just on the layer
   configuration files.
 * The layersScmOverrides and layersWhitelist keys were only
   appended when traversing the layers. This causes settings of
   unrelated layers to affect each other. Now the layer settings are
   handled strictly hierarchical.
 * It was possible to switch to the "import" SCM via layersScmOverrides.
   This is prevented now by always using the layers SCM schema.
 * There were no default environment white list entries. This was
   inconsistent to the regular white list handling.