layers: fix settings precedence
Layer configuration files on the command line should have the highest
precedence. If multiple files are passed, later ones should have a
higher precedence.

Settings in layer config.yaml files affect only their sub-layers. The
SCM overrides of a layer have a higher precedence than the SCM overrides
of a layer beneath it. This ensures that higher layers (or the command
line configuration files) can override and SCMs beneath it.