input: add stablePaths path configuration knob
The usage of stable paths (/bob/...) has traditionally been bound to the
usage of a sandbox. This makes sense insofar as these virtual paths
depend on the usage of mount namespaces.

The new "stablePaths" option allows to override the usage of virtual,
stable paths. By default, the existing heuristic is retained. But the
usage of stable paths can be now forced or prohibited too.