builder: add "slim sandbox" build mode
The slim sandbox uses the bob-namespace-sandbox to restrict access to
workspace paths. But in contrast to the existing "fat" sandbox, a slim
sandbox does not use a sandbox image. To gain at least a bit of
information hiding, the current working directory is hidden behind a
whiteout. This should usually hide all other workspaces.